+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Breaking
 * string interpolation is now `$" ... "`
 * `macroexpand` no longer expands `let-syntax`/`letrec-syntax` into `let`
+### Features
+* add R7RS promises with `delay-force`
 ### Bugfix
 * add `dfls` mnemonics for floating-point exponent
 * fix symbol syntax extensions (e.g. empty string interpolation `#""`) failing